requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.120.4
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
playwright==1.55.0
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
requests==2.32.5
requests-file==3.0.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.1
tldextract==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
gspread==6.1.4
google-auth==2.35.0
